+++ demos/chinook/src/main/java/is/codion/demos/chinook/server/ChinookAuthenticator.java
@@ -80,7 +80,7 @@ public final class ChinookAuthenticator implements Authenticator {
 	 * Handles logins from clients of this type
 	 */
 	@Override
-	public Optional<String> type() {
+	public Optional<String> clientType() {
 		return Optional.of("is.codion.framework.demos.chinook.ui.ChinookAppPanel");
 	}